+++ src/components/button-group/button-group.spec.ts
@@ -448,6 +448,34 @@ describe('Button Group', () => {
         expect(buttonGroup.selectedItems.length).to.equal(0);
       });
 
+      it('should clear the selection when passing an empty array to `selectedItems` property', () => {
+        buttons[0].selected = true;
+
+        expect(buttonGroup.selectedItems.length).to.equal(1);
+        expect(buttonGroup.selectedItems).to.have.same.members([
+          buttons[0].value,
+        ]);
+
+        buttonGroup.selectedItems = [];
+
+        expect(buttonGroup.selectedItems.length).to.equal(0);
+        expect(buttons[0].selected).to.be.false;
+      });
+
+      it('should clear the selection when passing falsy values to `selectedItems` property', () => {
+        buttons[0].selected = true;
+
+        expect(buttonGroup.selectedItems.length).to.equal(1);
+        expect(buttonGroup.selectedItems).to.have.same.members([
+          buttons[0].value,
+        ]);
+
+        buttonGroup.selectedItems = null as unknown as string[];
+
+        expect(buttonGroup.selectedItems.length).to.equal(0);
+        expect(buttons[0].selected).to.be.false;
+      });
+
       it('initial selection through child selection attribute has higher priority', async () => {
         // single mode
         buttonGroup = await createButtonGroupComponent(`